Add ApplicationHandler class
