+++ composer.json
@@ -18,7 +18,7 @@
     },
     "require": {
         "php": ">=5.5.0",
-        "laravel/framework": "~5.2.0",
+        "laravel/framework": "~5.2.7",
         "orchestra/database": "~3.2.0",
         "fzaninotto/faker": "~1.4",
         "symfony/css-selector": "2.8.*|3.0.*",
